+++ cmd/hidden_lake/messenger/internal/database/keys.go
@@ -6,8 +6,8 @@ import (
 
 const (
 	cKeySizeTemplate          = "database[%s].friends[%s].size"
+	cKeyMessageByUIDTemplate  = "database[%s].friends[%s].messages[uid=%x]"
 	cKeyMessageByEnumTemplate = "database[%s].friends[%s].messages[enum=%d]"
-	cKeyMessageByHashTemplate = "database[%s].friends[%s].messages[hash=%x]"
 )
 
 func getKeySize(pR IRelation) []byte {
@@ -27,11 +27,11 @@ func getKeyMessageByEnum(pR IRelation, pI uint64) []byte {
 	))
 }
 
-func getKeyMessageByHash(pR IRelation, pHash []byte) []byte {
+func getKeyMessageByUID(pR IRelation, pBlockUID [cBlockUIDSize]byte) []byte {
 	return []byte(fmt.Sprintf(
-		cKeyMessageByHashTemplate,
+		cKeyMessageByUIDTemplate,
 		pR.IAm().GetAddress().ToString(),
 		pR.Friend().GetAddress().ToString(),
-		pHash,
+		pBlockUID[:],
 	))
 }