+++ cmd/hidden_lake/messenger/internal/database/message.go
@@ -2,10 +2,16 @@ package database
 
 import (
 	"bytes"
-	"fmt"
 	"time"
 
-	"github.com/number571/go-peer/pkg/crypto/hashing"
+	"github.com/number571/go-peer/pkg/crypto/random"
+	"github.com/number571/go-peer/pkg/encoding"
+)
+
+const (
+	cIsIncomingSize = 1
+	cTimestampSize  = encoding.CSizeUint64
+	cBlockUIDSize   = 16
 )
 
 var (
@@ -14,38 +20,44 @@ var (
 
 type sMessage struct {
 	fIsIncoming bool
-	fSHA256UID  []byte
-	fTimestamp  string
+	fTimestamp  uint64
+	fBlockUID   [cBlockUIDSize]byte
 	fMessage    []byte
 }
 
-func NewMessage(pIsIncoming bool, pMessage []byte, pHashUID []byte) IMessage {
-	t := time.Now()
+func NewMessage(pIsIncoming bool, pMessage []byte) IMessage {
+	blockUID := [cBlockUIDSize]byte{}
+	copy(blockUID[:], random.NewStdPRNG().GetBytes(cBlockUIDSize))
+
 	return &sMessage{
 		fIsIncoming: pIsIncoming,
-		fSHA256UID:  pHashUID,
-		// 2022-12-08T02:43:24
-		fTimestamp: fmt.Sprintf("%04d-%02d-%02dT%02d:%02d:%02d",
-			t.Year(), t.Month(), t.Day(),
-			t.Hour(), t.Minute(), t.Second()),
-		fMessage: pMessage,
+		fBlockUID:   blockUID,
+		fTimestamp:  uint64(time.Now().Unix()),
+		fMessage:    pMessage,
 	}
 }
 
 func LoadMessage(pMsgBytes []byte) IMessage {
-	// 1(isIncoming) + 32(sha256) + 19(time)
-	if len(pMsgBytes) < 1+hashing.CSHA256Size+19 {
+	if len(pMsgBytes) < (cIsIncomingSize + cTimestampSize + cBlockUIDSize) {
 		return nil
 	}
+
 	isIncoming := false
 	if pMsgBytes[0] == 1 {
 		isIncoming = true
 	}
+
+	blockTimestamp := [cTimestampSize]byte{}
+	copy(blockTimestamp[:], pMsgBytes[cIsIncomingSize:cIsIncomingSize+cTimestampSize])
+
+	blockUID := [cBlockUIDSize]byte{}
+	copy(blockUID[:], pMsgBytes[cIsIncomingSize:cIsIncomingSize+cTimestampSize:cIsIncomingSize+cTimestampSize+cBlockUIDSize])
+
 	return &sMessage{
-		fIsIncoming: isIncoming,                                                          // 1 byte
-		fSHA256UID:  pMsgBytes[1 : hashing.CSHA256Size+1],                                // 32 bytes
-		fTimestamp:  string(pMsgBytes[hashing.CSHA256Size+1 : hashing.CSHA256Size+1+19]), // 19 bytes
-		fMessage:    pMsgBytes[hashing.CSHA256Size+1+19:],                                // n bytes
+		fIsIncoming: isIncoming,
+		fBlockUID:   blockUID,
+		fTimestamp:  encoding.BytesToUint64(blockTimestamp),
+		fMessage:    pMsgBytes[cIsIncomingSize+cTimestampSize+cBlockUIDSize:],
 	}
 }
 
@@ -53,8 +65,8 @@ func (p *sMessage) IsIncoming() bool {
 	return p.fIsIncoming
 }
 
-func (p *sMessage) GetSHA256UID() []byte {
-	return p.fSHA256UID
+func (p *sMessage) GetBlockUID() [cBlockUIDSize]byte {
+	return p.fBlockUID
 }
 
 func (p *sMessage) GetMessage() []byte {
@@ -62,7 +74,7 @@ func (p *sMessage) GetMessage() []byte {
 }
 
 func (p *sMessage) GetTimestamp() string {
-	return p.fTimestamp
+	return time.Unix(int64(p.fTimestamp), 0).Format("2006-01-02T15:04:05")
 }
 
 func (p *sMessage) ToBytes() []byte {
@@ -70,11 +82,12 @@ func (p *sMessage) ToBytes() []byte {
 	if p.fIsIncoming {
 		firstByte = byte(1)
 	}
+	blockTimestamp := encoding.Uint64ToBytes(p.fTimestamp)
 	return bytes.Join(
 		[][]byte{
 			{firstByte},
-			[]byte(p.fSHA256UID),
-			[]byte(p.fTimestamp),
+			blockTimestamp[:],
+			[]byte(p.fBlockUID[:]),
 			[]byte(p.fMessage),
 		},
 		[]byte{},